+++ libs/redpanda_qt_utils/qt_utils/utils.h
@@ -135,8 +135,10 @@ bool isInFolder(const QString& folderpath, const QString& filepath);
 QString includeTrailingPathDelimiter(const QString& path);
 QString excludeTrailingPathDelimiter(const QString& path);
 QString changeFileExt(const QString& filename, QString ext);
-
+QString getFilePath(const QString& folder, const QString& filename);
+QString getAbsoluteFilePath(const QString& folder, const QString& filename);
 QString localizePath(const QString& path);
+QString generateSubfolderPath(const QString& parentFolder, std::vector<QString> subfoldernames);
 
 QString extractRelativePath(const QString& base, const QString& dest);
 QStringList extractRelativePaths(const QString& base, const QStringList& destList);