+++ NEWS.md
@@ -1,5 +1,6 @@
 Red Panda C++ Version 3.3
   - enhancement: Auto indent contents in parenthesis.
+  - fix: Use "/" instead of "\" in the path of compile target. 
 
 Red Panda C++ Version 3.2